requirements.txt
qrcode[pil]
requests
Pillow
